# src/http/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Tracking": {
				"properties": {
					"orderNo": {
						"type": "string"
					},
					"tracking_number": {
						"type": "number",
						"format": "double"
					},
					"courier": {
						"type": "string"
					},
					"street": {
						"type": "string"
					},
					"zip_code": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"destination_country_iso3": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"articleNo": {
						"type": "string"
					},
					"articleImageUrl": {
						"type": "string"
					},
					"quantity": {
						"type": "string"
					},
					"product_name": {
						"type": "string"
					}
				},
				"required": [
					"orderNo",
					"tracking_number",
					"courier",
					"street",
					"zip_code",
					"city",
					"destination_country_iso3",
					"email",
					"articleNo",
					"articleImageUrl",
					"quantity",
					"product_name"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "generic-api",
		"description": "main api",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/orders": {
			"get": {
				"operationId": "GetOrders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Tracking"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/orders/{orderNo}": {
			"get": {
				"operationId": "GetOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Tracking"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "orderNo",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/orders/{orderNo}/articles": {
			"get": {
				"operationId": "GetTracking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Tracking"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "orderNo",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}